# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateContactRequest:
    properties:
      vcard_data:
        example: BEGIN:VCARD...
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.MoveContactRequest:
    properties:
      target_addressbook_id:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_domain_contact.Address:
    properties:
      city:
        type: string
      country:
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      type:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.Contact:
    properties:
      addressbook_id:
        description: UUID of the AddressBook
        type: string
      addresses:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Address'
        type: array
      birthday:
        description: Other
        type: string
      created_at:
        type: string
      emails:
        description: Lists
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Email'
        type: array
      etag:
        type: string
      family_name:
        type: string
      formatted_name:
        type: string
      given_name:
        type: string
      id:
        description: UUID of the AddressObject
        type: string
      middle_name:
        type: string
      nickname:
        type: string
      notes:
        type: string
      organization:
        description: Organization
        type: string
      phones:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Phone'
        type: array
      photo_url:
        description: Constructed URL for the photo
        type: string
      prefix:
        description: Name components
        type: string
      suffix:
        type: string
      title:
        type: string
      uid:
        description: vCard UID
        type: string
      updated_at:
        type: string
      urls:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.URL'
        type: array
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.Email:
    properties:
      primary:
        type: boolean
      type:
        type: string
      value:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.Phone:
    properties:
      primary:
        type: boolean
      type:
        type: string
      value:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.URL:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_usecase_contact.ListOutput:
    properties:
      contacts:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_jherrma_caldav-server_internal_usecase_contact.SearchOutput:
    properties:
      contacts:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        type: array
      count:
        type: integer
      query:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_usecase_contact.UpdateInput:
    properties:
      addresses:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Address'
        type: array
      birthday:
        type: string
      emails:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Email'
        type: array
      family_name:
        type: string
      formatted_name:
        type: string
      given_name:
        type: string
      middle_name:
        type: string
      nickname:
        type: string
      notes:
        type: string
      organization:
        type: string
      phones:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Phone'
        type: array
      prefix:
        type: string
      suffix:
        type: string
      title:
        type: string
      urls:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.URL'
        type: array
    type: object
...
  internal_adapter_http.ErrorResponseBody:
    description: Error response body
    properties:
      error:
        example: error_code
        type: string
      message:
        example: Human readable message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HTTP Basic Authentication for DAV endpoints
  title: CalDAV/CardDAV Server API
  version: "1.0"
paths:
...
  /addressbooks/{addressbook_id}/contacts:
    get:
      description: Get contacts from address book
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Limit (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Sort field (default name)
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_usecase_contact.ListOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: List contacts
      tags:
      - Contacts
    post:
      consumes:
      - application/json
      description: Create a new contact
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact details
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create contact
      tags:
      - Contacts
  /addressbooks/{addressbook_id}/contacts/{contact_id}:
    delete:
      description: Delete a contact
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact UUID
        in: path
        name: contact_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete contact
      tags:
      - Contacts
    get:
      description: Get contact by ID
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact UUID
        in: path
        name: contact_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Get contact
      tags:
      - Contacts
    put:
      consumes:
      - application/json
      description: Update contact details
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact UUID
        in: path
        name: contact_id
        required: true
        type: string
      - description: Contact updates
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_usecase_contact.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Update contact
      tags:
      - Contacts
  /addressbooks/{addressbook_id}/contacts/{contact_id}/photo:
    delete:
      description: Remove photo from contact
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact UUID
        in: path
        name: contact_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete contact photo
      tags:
      - Contacts
    get:
      description: Get contact's photo content
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact UUID
        in: path
        name: contact_id
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - image/gif
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Get contact photo
      tags:
      - Contacts
    put:
      consumes:
      - image/jpeg
      - image/png
      - image/gif
      description: Upload a photo for a contact (JPEG, PNG, GIF)
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact UUID
        in: path
        name: contact_id
        required: true
        type: string
      - description: Photo data
        in: body
        name: file
        required: true
        schema:
          items:
            type: integer
          type: array
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "413":
          description: Photo too large
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "415":
          description: Unsupported file type
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Upload contact photo
      tags:
      - Contacts
...
  /addressbooks/{id}/contacts:
    post:
      consumes:
      - application/json
      description: Create a new contact in address book
      parameters:
      - description: Address Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contact VCard
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'Ref: domain.AddressObject'
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create contact
      tags:
      - Contacts
...
  /contacts/{contact_id}/move:
    post:
      consumes:
      - application/json
      description: Move contact to another address book
      parameters:
      - description: Contact UUID
        in: path
        name: contact_id
        required: true
        type: string
      - description: Target address book
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.MoveContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Move contact
      tags:
      - Contacts
  /contacts/search:
    get:
      description: Search for contacts by query
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Address Book ID filter
        in: query
        name: addressbook_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_usecase_contact.SearchOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Search contacts
      tags:
      - Contacts
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...